moved followStatus query to within the FollowButton component
