+++ package.json
@@ -14,7 +14,6 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
-    "@types/node": "^16.18.11",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.10",
     "axios": "^1.3.2",
@@ -30,6 +29,7 @@
     "web-vitals": "^2.1.4"
   },
   "devDependencies": {
+    "@types/node": "^20.11.10",
     "@vitejs/plugin-react": "^4.1.1",
     "sass": "^1.69.5",
     "vite": "^3.2.2",
