+++ src/components/NavBar/NavBar.tsx
@@ -56,7 +56,7 @@ const NavBar = () => {
         <Box sx={styles.navList}>
           <Avatar
             alt="logo"
-            src="https://www.iconpacks.net/icons/2/free-twitter-logo-icon-2429-thumb.png"
+            src={`/chirp-logo-transparent.png`}
             sx={styles.logo}
           />
           <List component="nav">
