+++ src/pages/Profile.tsx
@@ -103,7 +103,6 @@ const Profile = () => {
   const navigate = useNavigate();
   const { username } = useParams();
   const [value, setValue] = useState("one");
-  const [followStatus, setFollowStatus] = useState(false);
   const user = useAppSelector((state) => state.user);
   const [profileContents, setProfileContents] = useState<ProfileContent>({
     postCount: 0,
@@ -114,25 +113,6 @@ const Profile = () => {
     followerCount: 0,
     followingCount: 0,
   });
-
-  useEffect(() => {
-    const fetchFollowStatus = async () => {
-      const result = await axios.get(
-        "http://localhost:3001/api/profile/getFollowStatus",
-        {
-          params: {
-            userId: user.userId,
-            username,
-          },
-        }
-      );
-      setFollowStatus(result.data.followStatus);
-    };
-
-    if (profileContents.username !== username) {
-      fetchFollowStatus();
-    }
-  }, [profileContents.username, username, user.userId]);
 
   useEffect(() => {
     const fetchProfileContents = async () => {
@@ -155,11 +135,7 @@ const Profile = () => {
     };
     fetchProfileContents();
     window.scrollTo(0, 0);
-  }, [value, username, followStatus]);
-
-  const fetchFollowStatus = (followStatus: boolean) => {
-    setFollowStatus(followStatus);
-  };
+  }, [value, username]);
 
   return (
     <Layout
@@ -193,10 +169,7 @@ const Profile = () => {
                     Edit Profile
                   </Button>
                 ) : (
-                  <FollowButton
-                    username={username}
-                    fetchFollowStatus={fetchFollowStatus}
-                  />
+                  <FollowButton username={username} />
                 )}
               </Box>
               <Box sx={styles.nameContainer}>
