+++ vite.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
   // comment this out if that isn't relevant for your project
   build: {
     outDir: "build",
+    chunkSizeWarningLimit: 1000,
   },
   server: {
     port: 3000,
